third_party/blink/web_tests/platform/mac-mac10.10/wpt_internal/webxr/xrSession_environmentBlendMode.https-expected.txt

This is a testharness.js-based test.
PASS environmentBlendMode is correct for a VR device in immersive
PASS environmentBlendMode is correct for a VR device in non-immersive
Harness: the test ran to completion.
